Move import hardware set functionality to menu
