+++ tests/gui/hardware_set/conftest.py
@@ -6,7 +6,9 @@ from pathlib import Path
 import pytest
 
 from frog.gui.hardware_set.device import OpenDeviceArgs
-from frog.gui.hardware_set.hardware_set import HardwareSet
+from frog.gui.hardware_set.hardware_set import (
+    HardwareSet,
+)
 
 _HW_SETS = (
     HardwareSet(
